feat: Add executable and main library for Tangome project
This commit introduces an executable that specifies the Ruby version and the app path. In addition, it adds the main library for the 'Tangome' project including logic for dictionary search, edit, addition, and displaying a list or a single word. It also creates a new class file that has the dictionary logic and web search feature.